# Optimize OStream implementation