When writing annotation heavy code, we spend 80% of our time in this routine:

<img width="1281" alt="screen shot 2019-02-21 at 1 53 13 pm" src="https://user-images.githubusercontent.com/9171254/53193878-1901b700-35e0-11e9-98f6-9e1b29cfbec4.png">

This P.R makes that code-path 10x faster

Before
```
goos: darwin
goarch: amd64
pkg: github.com/m3db/m3/src/dbnode/encoding
BenchmarkWriteBytes-8   	  100000	     22507 ns/op	      20 B/op	       0 allocs/op
PASS
ok  	github.com/m3db/m3/src/dbnode/encoding	2.495s
Success: Benchmarks passed.
```

After
```
goos: darwin
goarch: amd64
pkg: github.com/m3db/m3/src/dbnode/encoding
BenchmarkWriteBytes-8   	 1000000	      2226 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/m3db/m3/src/dbnode/encoding	2.271s
```